Move phdi.fhir.transport module code to http.py
